Remove magic comma logic from can_omit_invisible_parens
It was causing stability issues because the first pass
could cause a "magic trailing comma" to appear, meaning
that the second pass might get a different result. It's
not critical.

black-primer previously crashed on several projects due to instability, but now passes.
The preexisting disabled tests can now be enabled. Added a new
test for a case that came up in my testing.

Some things format differently (with extra parens)