The pants runner script.
This is a standard script that is downloaded from
https://static.pantsbuild.org/setup/pants per the
Pants installation instructions. This script takes
care of installing Pants at the right version, and
then executing it.

This commit also includes a pants_from_sources script,
which is useful for running Pants directly from its own
sources, on the code in this repo. This, again, is a
standard script commonly used for this purpose.

It also includes .gitignore updates relevant to Pants
byproducts.