The Pants BUILD files.
These contain metadata that describe the structure of the
repo, and things that can be built from it.

This commit also includes requirements.txt files for
the BUILD files to wrap (this is the standard way to
describe 3rdparty requirements to Pants), and a macro
that allows one of our BUILD files to cut down on a lot
of boilerplate.

Unlike other similar systems, Pants BUILD files typically
don't require explicit lists of dependencies, as Pants infers
those via static analysis. However occasionally some manual
tweaks are required, such as for dynamic runtime dependencies,
or data dependencies.