Some minor code changes to accommodate Pants.
These are worth paying particular attention to in review.

They are mostly around the difference in how setuptools_scm
is invoked to generate the Black version string. Pants uses its
codegen mechanism for this, via the vcs_version target type.
This means that the version data is generated on the fly, whenever
an import requests it.

This also adds a few pragmas to tell Pants not to infer dependencies
that are provided externally.