Fix the handling of `# fmt: skip` when it's at a colon line.
When the Leaf node with `# fmt: skip` is a NEWLINE inside a `suite` Node, the nodes to ignore should be from the siblings of the parent `suite` Node.

There is a also a special case for the ASYNC token, where it expands to the grandparent Node where the ASYNC token is.

This fixes #2646, #3126, #2680, #2421, #2339, #2138.