Fix test that passes on main anyway
Black searches parent directories if the common base of the given
sources doesn't contain a "project root marker" so simply setting
--stdin-filename to `../whatever.py` won't cause any problems. Instead
let's set --stdin-filename to `../src/a.py` as `src` has its own
pyproject.toml file. Currently on main, the detected root will be the
parent directory of `src` but with this PR it'll be `src` as expected.