Always call freeze_support() if sys.frozen is True
This shouldn't be necessary, but in practice macOS also requires
freeze_support() so let's play it safe and always call it when
sys.frozen is truthy. Only a tiny sliver of users will be using these
PyInstaller executables and the start-up performance is already terrible
as the executable unzips itself on the fly anyway.