Fix printing of emoji on Windows when stdout is redirected
This is most apparent when black is run via pre-commit which does
`subprocess.Popen(stdout=PIPE)`.

One option around this would be to instruct users to set `PYTHONUTF8=1`,
but that is not a very good approach, as basically everything supports
UTF-8 these days.

This fix was inspired by pycln and https://github.com/hadialqattan/pycln/issues/54