Create Decisiontree
This model is a decision tree classifier that predicts whether a person will buy a car or not based on their age and income.In this example, the X array contains the input data, which consists of age and income values for eight different people. The Y array contains the corresponding output values, which indicate whether each person bought a car or not.

The decision tree classifier is created using the DecisionTreeClassifier class from scikit-learn. The fit method is then called to train the model using the input and output data. Finally, the predict method is used to make a prediction for a person with age 45 and income 110000. The output of the program should be 'Yes', indicating that the person is likely to buy a car.

Note that this is just a simple example to demonstrate how machine learning models can be implemented in Python using scikit-learn. In practice, more complex models and larger datasets would be used.