Edit test to make sure a delimiter split is performed
This test makes sure a trailing comma is added when there is a comment
in the brackets. However, a delimiter split is no longer performed in
these cases, so there is no need to add a trailing comma. This commit
makes the lines longer so a delimiter split is performed, and the
behaviour with the trailing comma can be tested.