Add heuristic to use correct sibling/preceding leaf
In cases where 'fmt:skip' belonged to a closing bracket, prev_sibling would give us the opening bracket instead of the previous element inside the brackets. This is desirable if the whole thing is on the same line (as that is what fmt:skip targets), but in other cases we actually want to start processing the preceding leaf instead.