Remove redundant parentheses in `case` statement `if` guards:
A follow up to #4024 but for `if` guards in `case` statements. I noticed this
when #4024 was made stable, and noticed I had some code that had extra parens
around the `if` guard.

There might be some edge cases I am not thinking of, but all tests are passing.
I put these new changes under the `remove_redundant_guard_parens` preview name.