Fix schema generation error caused by new click version (#4750)
* Fix schema generation error caused by new click version
Not sure on the policy for updating versions, so to cope with the new click 8.3.0 version I used some code from #4577 for version conditional behavior. When >= 8.3.0, unset defaults are now the special UNSET sentinal instead of None.
Hopefully nothing breaks on the click._utils usage since it's an _ module, might have to find a different solution if it does.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add ignore for mypy error

* Actually fix mypy errors with ignore

* Add explanation for ignores

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>